% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_NEI.R
\name{compile_NEI}
\alias{compile_NEI}
\title{Compile NEI data}
\usage{
compile_NEI(data_folder)
}
\arguments{
\item{data_folder}{File path to a local folder containing .csv files of EPA NEI Original data to process + compile.}
}
\value{
A dataframe of compiled NEI data.
}
\description{
Compile NEI data
}
