% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_NEI.R
\name{map_NEI}
\alias{map_NEI}
\title{Map NEI data}
\usage{
map_NEI(data, sector)
}
\arguments{
\item{data}{.csv file of compiled NEI data.}

\item{sector}{string input of the pollutant sector of interest.}
}
\value{
A ggplot object of spatially modeled Virginia pollutant sources.
}
\description{
Map NEI data
}
