% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_NEI.R
\name{process_NEI}
\alias{process_NEI}
\title{Process NEI data}
\usage{
process_NEI(df)
}
\arguments{
\item{df}{.csv file containing EPA NEI Original data.}
}
\value{
A dataframe of cleaned NEI data
}
\description{
Process NEI data
}
